Suicides in India: What Data Shows?
Introduction:
Suicide or ending one’s own life is a tragic event with strong emotional repercussions for its survivors and for families of its victims.

India accounts for a growing share of the world’s suicides.

The nation accounts for over a third of the world's annual female suicides and nearly a fourth of male suicides, a significant rise in global share from 1990.

Each suicide is a personal tragedy that prematurely takes the life of an individual and has a continuing ripple effect, dramatically affecting the lives of families, friends and communities. Every year, more than 1,00,000 people commit suicide in our country. There are various causes of suicides like professional/career problems, discrimination,sense of isolation, abuse, violence, family problems, mental disorders, addiction to alcohol, financial loss, chronic pain etc.

Objective and Scope of the Project
Objective :
The primary objective of the study are:

Study suicide rates for various locations in India.
Identify the major causes that contribute to the most of the suicide attacks in India.
Analyse Suicides attacks on the basis of means and status of victims.
Scope :
This data set contains yearly suicide detail of all the states/u.t of India by various parameters from 2001 to 2012.

Tools used for Visualisation : Tableau and Python Visualisation Libraries